feat(front: load user): loadUser関数をhook化、userとroasterのグローバルステートを更新する
