+++ api/app/controllers/api/v1/auth/sessions_controller.rb
@@ -4,6 +4,7 @@ class Api::V1::Auth::SessionsController < DeviseTokenAuth::SessionsController
 
   def index
     @current_user = current_api_v1_user
+    @current_roaster = current_api_v1_user.roaster if current_api_v1_user
     render formats: :json
   end
 end
