+++ front/src/features/roasters/hooks/useCurrentRoaster.ts
@@ -1,5 +1,3 @@
-import { useState } from 'react'
-
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 
 import { currentRoasterState } from '@/features/roasters/stores/currentRoasterState'
@@ -9,8 +7,6 @@ import type { Roaster } from '@/features/roasters/types'
 import type { SetterOrUpdater } from 'recoil'
 
 export const useCurrentRoaster = () => {
-  const [loading, setLoading] = useState(false)
-
   // Recoilでグローバルステートを定義
   const currentRoaster = useRecoilValue(currentRoasterState) // Getterを定義
   const setCurrentRoaster: SetterOrUpdater<Roaster | null> = useSetRecoilState(currentRoasterState) // Setter, Updaterを定義
@@ -19,5 +15,5 @@ export const useCurrentRoaster = () => {
   const isRoaster = useRecoilValue(isRoasterState)
   const setIsRoaster = useSetRecoilState(isRoasterState)
 
-  return { currentRoaster, setCurrentRoaster, isRoaster, setIsRoaster, loading }
+  return { currentRoaster, setCurrentRoaster, isRoaster, setIsRoaster }
 }
