+++ front/src/features/users/components/organisms/UserUpdateForm.tsx
@@ -7,7 +7,7 @@ import { useForm } from 'react-hook-form'
 
 import { PrimaryButton } from '@/components/Elements/Button'
 import { ImagePreview } from '@/components/Form'
-import { useAuth, useSignedInUser } from '@/features/auth'
+import { useAuth, useLoadUser } from '@/features/auth'
 import { updateUser } from '@/features/users/api/updateUser'
 import { EmailInput } from '@/features/users/components/molecules/EmailInput'
 import { PrefectureSelect } from '@/features/users/components/molecules/PrefectureSelect'
@@ -36,8 +36,9 @@ type UserUpdateDate = UserUpdateParams & {
 export const UserUpdateForm: FC<Props> = (props) => {
   const { user, setIsError } = props
   const { authHeaders } = useAuth()
-  const { loadUser } = useSignedInUser()
   const navigate = useNavigate()
+
+  const { loadUser } = useLoadUser()
   const { showMessage } = useMessage()
   const { setErrorNotifications } = useErrorNotification()
